commit by Satyam (controller added)

+++ fms-register-and-login-module/src/main/java/com/example/application/utility/Encryption.java
@@ -4,7 +4,7 @@ public class Encryption {
 
 	public static String passwordEncryptor(String password) {
 		int key = 7;
-		int encryptedNumber = 0;
+		long encryptedNumber = 0;
 		for(int i = 0; i < password.length(); i++) {
 			encryptedNumber += (int)password.charAt(i) * key;
 		}
